bugix from manual testing
