+++ src/core/classes/Form.class.ts
@@ -77,7 +77,9 @@ class Form {
     this.securityCode = new Element();
     this.animatedCard = new Element();
 
-    this.cardNumber.create(Selectors.CARD_NUMBER_COMPONENT_NAME, this.styles);
+    this.cardNumber.create(Selectors.CARD_NUMBER_COMPONENT_NAME, this.styles, {
+      origin: this.origin
+    });
     this.cardNumberMounted = this.cardNumber.mount(Selectors.CARD_NUMBER_IFRAME);
     this.elementsToRegister.push(this.cardNumberMounted);
